requirements.txt
asgiref==3.8.1
boto3==1.34.144
botocore==1.34.144
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
jmespath==1.0.1
mysql-client==0.0.1
pi==0.1.2
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary>=2.8
psycopg-pool==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==1.26.19
